moved type Comment to commentItem
